move init_lmcache_engine() to v1 adapter
