+++ lmcache/integration/vllm/vllm_adapter.py
@@ -36,20 +36,13 @@ from vllm.config import (
     SchedulerConfig,
 )
 from vllm.sequence import IntermediateTensors
-from vllm.utils import cdiv, get_kv_cache_torch_dtype, round_down
+from vllm.utils import cdiv, round_down
 
 # First Party
-from lmcache.config import LMCacheEngineMetadata
-from lmcache.integration.vllm.utils import ENGINE_NAME, lmcache_get_config
+from lmcache.integration.vllm.utils import ENGINE_NAME
 from lmcache.logging import init_logger
 from lmcache.utils import _lmcache_nvtx_annotate
-from lmcache.v1.cache_engine import LMCacheEngine, LMCacheEngineBuilder
-from lmcache.v1.config import LMCacheEngineConfig
-from lmcache.v1.gpu_connector import (
-    VLLMBufferLayerwiseGPUConnector,
-    VLLMPagedMemGPUConnectorV2,
-    VLLMPagedMemLayerwiseGPUConnector,
-)
+from lmcache.v1.cache_engine import LMCacheEngineBuilder
 
 # FIXME(Jiayi): temporarily comment this out
 # from lmcache_vllm.blend_adapter import remove_request_id_indices
@@ -84,137 +77,6 @@ class RetrieveStatus(Enum):
     # (3) prefix_prefill
     CHUNK_PREFILL = 2  # not last chunk
     NONE = 4
-
-
-def need_gpu_interm_buffer(lmcache_config: LMCacheEngineConfig):
-    if lmcache_config.enable_nixl:
-        return False
-    else:
-        return True
-
-
-def init_lmcache_engine(
-    model_config: ModelConfig,
-    parallel_config: ParallelConfig,
-    cache_config: CacheConfig,
-    scheduler_config: SchedulerConfig,
-) -> Optional[LMCacheEngine]:
-    """Initialize the LMCache engine by the given model config and parallel
-    config. This function will check the environment variable
-    `LMCACHE_CONFIG_FILE` to load the configuration file. If that environment
-    variable is not set, this function will return None.
-
-    :param model_config: The model configuration in vLLM.
-    :type model_config: ModelConfig
-    :param parallel_config: The parallel configuration in vLLM.
-    :type parallel_config: ParallelConfig
-    :param cache_config: The KV cache configuration in vLLM.
-    :type cache_config: CacheConfig
-    :param scheduler_config: The scheduler configuration in vLLM.
-    :type scheduler_config: SchedulerConfig
-
-    :return: The initialized LMCache engine or None (if the environment variable
-        `LMCACHE_CONFIG_FILE` is not set).
-    :rtype: Optional[LMCacheEngine]
-    """
-    if LMCacheEngineBuilder.get(ENGINE_NAME) is not None:
-        return None
-
-    global VLLM_CACHE_CONFIG
-    global VLLM_PARALLEL_CONFIG
-    global VLLM_MODEL_CONFIG
-    global VLLM_SCHEDULER_CONFIG
-    VLLM_CACHE_CONFIG = cache_config
-    VLLM_PARALLEL_CONFIG = parallel_config
-    VLLM_MODEL_CONFIG = model_config
-    VLLM_SCHEDULER_CONFIG = scheduler_config
-
-    config = lmcache_get_config()
-    assert isinstance(config, LMCacheEngineConfig), (
-        "LMCache v1 configuration is should be passed."
-    )
-
-    kv_dtype = get_kv_cache_torch_dtype(cache_config.cache_dtype, model_config.dtype)
-
-    use_mla = False
-    if (
-        hasattr(model_config, "use_mla")
-        and isinstance(model_config.use_mla, bool)
-        and model_config.use_mla
-    ):
-        use_mla = True
-
-    if use_mla and (config.remote_serde != "naive" and config.remote_serde is not None):
-        raise ValueError("MLA only works with naive serde mode..")
-
-    # construct kv shape (for mem pool)
-    num_layer = model_config.get_num_layers(parallel_config)
-    chunk_size = config.chunk_size
-    num_kv_head = model_config.get_num_kv_heads(parallel_config)
-    head_size = model_config.get_head_size()
-    kv_shape = (num_layer, 1 if use_mla else 2, chunk_size, num_kv_head, head_size)
-    logger.info(f"use mla: {use_mla}, kv shape: {kv_shape}")
-
-    # Change current device.
-    torch.cuda.device(parallel_config.rank)
-    device = torch.device(f"cuda:{parallel_config.rank}")
-    metadata = LMCacheEngineMetadata(
-        model_config.model,
-        parallel_config.world_size,
-        parallel_config.rank,
-        "vllm",
-        kv_dtype,
-        kv_shape,
-        use_mla,
-    )
-
-    use_gpu = need_gpu_interm_buffer(config)
-    vllm_gpu_connector: Union[
-        VLLMBufferLayerwiseGPUConnector,
-        VLLMPagedMemGPUConnectorV2,
-        VLLMPagedMemLayerwiseGPUConnector,
-    ]
-
-    if use_mla and config.use_layerwise:
-        raise ValueError("layerwise MLA connector is not supported yet")
-
-    # When use_mla is True, num_kv_head is 1
-    hidden_dim_size = num_kv_head * head_size
-    if config.use_layerwise:
-        if config.enable_blending:
-            # Use layerwise connector for blending
-            vllm_gpu_connector = VLLMBufferLayerwiseGPUConnector(
-                hidden_dim_size,
-                num_layer,
-                use_gpu=use_gpu,
-                chunk_size=chunk_size,
-                dtype=kv_dtype,
-                device=device,
-            )
-        else:
-            vllm_gpu_connector = VLLMPagedMemLayerwiseGPUConnector(
-                hidden_dim_size,
-                num_layer,
-                use_gpu=use_gpu,
-                chunk_size=chunk_size,
-                dtype=kv_dtype,
-                device=device,
-            )
-    else:
-        vllm_gpu_connector = VLLMPagedMemGPUConnectorV2(
-            hidden_dim_size,
-            num_layer,
-            use_gpu=use_gpu,
-            chunk_size=chunk_size,
-            dtype=kv_dtype,
-            device=device,
-            use_mla=use_mla,
-        )
-    engine = LMCacheEngineBuilder.get_or_create(
-        ENGINE_NAME, config, metadata, vllm_gpu_connector
-    )
-
-    return engine
 
 
 def broadcast_seq_group_list(
